Creating a note_rowcount has been discredited and temporarilly removed.
The use of note_rowcount will change SQL%ROWCOUNT, if tested on the next line of code.
Potentially also the insertion of any logger commands between execution of SQL and the testing of SQL%ROWCOUNT
Given this, may be necessary for the AOP_PROCESSOR to created WARNINGS when SQL%ROWCOUNT tests are found in the code,
to ensure the Developer is aware of the issue.

Modified   database/LOGGER/Packages/aop_processor.pkb
--MOD PROCEDURE AOP_block - commented out note_rowcount
Modified   database/LOGGER/UnitTest/aop_test.pkb
Modified   database/LOGGER/UnitTest/aop_test.pks
--NEW FUNCTION test5 - tests for note_rowcount