Fixed Apex Context with APP_USER of nobody
The Apex Sessions count was not matching the sessions in Session List.
The problem was due to my SM_APEX_CONTEXT records being created with a user of "nobody", before the person finished logging in.
I've changed this so that the context records are updated to the real user, as soon as they are known.

Modified   database/LOGGER/Packages/sm_logger.pkb
--MOD procedure write_apex_context
--Update the app_user on any existing context still showing "nobody".